Solution to Calculate Average Temperature of a Week

+++ Calculate Average Temperature of a Week@@67877d61dc107486ad5c274d/code.c
@@ -1,14 +1,14 @@
 #include <stdio.h>
 
 struct week{
-    int temp;
+    float temp;
 };
 
 int main() {
     struct week e[7];
     float totalTemp=0 , average;
     for(int i=0;i<7;i++){
-        scanf("%d",&e[i].temp);
+        scanf("%f",&e[i].temp);
 
         totalTemp += e[i].temp;
     }
